@@ -4,20 +4,18 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
     "just-shuffle": "^4.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.5.3",
     "react-scripts": "5.0.0",
-    "recoil": "^0.6.1",
-    "typescript": "^4.6.2",
-    "web-vitals": "^2.1.4"
+    "recoil": "^0.6.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -43,5 +41,8 @@
       "last 1 safari version"
     ]
   },
-  "packageManager": "npm@11.2.0+sha512.3dc9c50ba813a3d54393155a435fe66404b72685ab0e3008f9ae9ed8d81f6104860f07ed2656dd5748c1322d95f3140fa9b19c59a6bba7750fd12285f81866da"
+  "packageManager": "npm@11.2.0+sha512.3dc9c50ba813a3d54393155a435fe66404b72685ab0e3008f9ae9ed8d81f6104860f07ed2656dd5748c1322d95f3140fa9b19c59a6bba7750fd12285f81866da",
+  "devDependencies": {
+    "typescript": "^5.8.3"
+  }
 }
