@@ -1,15 +1,13 @@
-import './index.css';
+import './index.css'
 
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+ReactDOM
+	.createRoot(document.getElementById('root') as HTMLElement)
+	.render(
+		<React.StrictMode>			
+			<App />			
+		</React.StrictMode>	
 );
-
-reportWebVitals();
