@@ -1,22 +1,22 @@
+import { BrowserRouter, Route, Routes } from 'react-router';
+import { RecoilRoot } from 'recoil';
+
 import Header from './components/Header/Header';
+import Form from './components/Form';
 
 function App() {
-  return (
-   <div className="App">
-      <Header />
-      <p>
-        Edit <code>src/App.tsx</code> and save to reload.
-      </p>
-      <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </div>
-  )
+    return (        
+        <>
+            <Header />
+            <BrowserRouter>
+                <RecoilRoot>
+                    <Routes>
+                        <Route path="/" element={<Form />} />
+                    </Routes>
+                </RecoilRoot>                           
+            </BrowserRouter>
+        </>
+    )
 }
 
-export default App;
+export default App
